requirements.txt
bertopic==0.17.3
fastapi==0.116.1
hdbscan==0.8.40
langdetect==1.0.9
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
pandas==2.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
scikit-learn==1.7.2
sentence-transformers==5.1.0
streamlit==1.49.1
torch==2.8.0
transformers==4.56.0
umap-learn==0.5.4
uvicorn==0.35.0